Moved Stash and BufferFiller classes into their own files
